fix: remove old validator workflow
